LayoutTests/imported/w3c/web-platform-tests/css/css-color/parsing/color-valid-hsl-expected.txt


PASS e.style['color'] = "hsl(120 30% 50%)" should set the property value
PASS e.style['color'] = "hsl(120 30% 50% / 0.5)" should set the property value
PASS e.style['color'] = "hsl(none none none)" should set the property value
PASS e.style['color'] = "hsl(0 0% 0%)" should set the property value
PASS e.style['color'] = "hsl(none none none / none)" should set the property value
PASS e.style['color'] = "hsl(0 0% 0% / 0)" should set the property value
PASS e.style['color'] = "hsla(none none none)" should set the property value
PASS e.style['color'] = "hsla(0 0% 0%)" should set the property value
PASS e.style['color'] = "hsla(none none none / none)" should set the property value
PASS e.style['color'] = "hsla(0 0% 0% / 0)" should set the property value
PASS e.style['color'] = "hsl(120 none none)" should set the property value
PASS e.style['color'] = "hsl(120 0% 0%)" should set the property value
PASS e.style['color'] = "hsl(120 80% none)" should set the property value
PASS e.style['color'] = "hsl(120 80% 0%)" should set the property value
PASS e.style['color'] = "hsl(120 none 50%)" should set the property value
PASS e.style['color'] = "hsl(120 0% 50%)" should set the property value
PASS e.style['color'] = "hsl(120 100% 50% / none)" should set the property value
PASS e.style['color'] = "hsl(120 100% 50% / 0)" should set the property value
PASS e.style['color'] = "hsl(none 100% 50%)" should set the property value
PASS e.style['color'] = "hsl(0 100% 50%)" should set the property value
FAIL e.style['color'] = "hsl(120 30 50)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hsl(120 30 50 / 0.5)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hsl(120 30% 50)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hsl(120 30% 50 / 0.5)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hsl(120 30 50%)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hsl(120 30 50% / 0.5)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hsl(120 none 50)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hsl(120 none 50 / 0.5)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hsl(120 30 none)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hsl(120 30 none / 0.5)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hsl(120 30 50 / none)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hsl(calc(infinity) 100% 50%)" should set the property value assert_equals: serialization should be canonical expected "rgb(255, 0, 0)" but got "rgb(0, 0, 0)"
FAIL e.style['color'] = "hsl(calc(-infinity) 100% 50%)" should set the property value assert_equals: serialization should be canonical expected "rgb(255, 0, 0)" but got "rgb(0, 0, 0)"
PASS e.style['color'] = "hsl(calc(0 / 0) 100% 50%)" should set the property value
PASS e.style['color'] = "hsl(90 50% 50% / calc(infinity))" should set the property value
PASS e.style['color'] = "hsl(90 50% 50% / calc(-infinity))" should set the property value
PASS e.style['color'] = "hsl(90 50% 50% / calc(0 / 0))" should set the property value
